Fix for 1.9.2.0
Version 1.9.2.0 no longer contains arweave-server.  In order to properly locate the arweave directory to pull log files dynamically, please ensure this script (arweave-monitor.py) is placed in the root of your arweave directory.  The script will then look for it self and use that location to find the logs.  This way, when Arweave updates their code again and changes their folder structure and/or start up commands, the script will still work.